seed

+++ database/migrations/2025_06_05_064619_create_merchants_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('merchant_name');
             $table->string('phone_number')->unique();
-            $table->unsignedInteger('id_account');
+            $table->unsignedBigInteger('id_account');
             $table->timestamps();
 
             $table->foreign('id_account')->on('accounts')->references('id');
